:recycle: Refs #547 - Get working again after rebase
