+++ src/openforms/js/components/admin/form_design/RegistrationFields.js
@@ -7,12 +7,14 @@ import FormRow from '../forms/FormRow';
 import Fieldset from '../forms/Fieldset';
 import Select from "../forms/Select";
 import {TextInput} from '../forms/Inputs';
+import Form from "@rjsf/core";
 
 
 const RegistrationFields = ({
     backends=[],
     selectedBackend='',
     backendOptions={},
+    backendOptionsForms={},
     onChange,
 }) => {
     const backendChoices = backends.map( backend => [backend.id, backend.label]);
@@ -29,6 +31,10 @@ const RegistrationFields = ({
         }
         const fakeEvent = {target: {name, value: parsedValue}};
         onChange(fakeEvent);
+    };
+
+    const onRegistrationBackendOptionsFormChange = ({ formData }) => {
+        onChange({target: {name: 'form.registrationBackendOptions', value: formData}});
     };
 
     return (
@@ -55,6 +61,23 @@ const RegistrationFields = ({
                         value={JSON.stringify(backendOptions || {})}
                         onChange={onBackendOptionsChange}
                         maxLength="1000" />
+                </Field>
+            </FormRow>
+            <FormRow>
+                <Field
+                    name="form.registrationBackendOptionsForm"
+                    label="Registration Backend Options Form"
+                >
+                    {backendOptionsForms[selectedBackend] ?
+                        <Form
+                            schema={backendOptionsForms[selectedBackend]}
+                            formData={backendOptions}
+                            onChange={onRegistrationBackendOptionsFormChange}
+                            children={true}
+                        />
+                        :
+                        <div> </div>
+                    }
                 </Field>
             </FormRow>
         </Fieldset>
